+++ src/Symfony/Component/DomCrawler/CHANGELOG.md
@@ -1,6 +1,19 @@
 CHANGELOG
 =========
 
+3.0.0
+-----
+
+* [BC BREAK] `Form` now extends `AbstractUriElement` instead of `Link`.
+* [BC BREAK] The `node` and `method` properties of `AbstractUriElement` are now private (they already have public getters).
+* [BC BREAK] Since `node` is private, `setNode` in child classes can't set the `node` property directly, thus `setNode` have been renamed to `findNode` and now returns the node instead of setting the property.
+
+2.7.0
+-----
+
+* All the URI parsing logic have been abstracted in the `AbstractUriElement` class. The `Link` class is now a child of `AbstractUriElement` which implements the new `UriElementInterface`, describing the common `getNode`, `getMethod` and `getUri` methods.
+* Added an `Image` class to crawl images and parse their `src` attribute, and `selectImage`, `image`, `images` methods in `Crawler`, the image version of the equivalent `link` methods.
+
 2.5.0
 -----